+++ src/essarion_build/__init__.py
@@ -83,7 +83,7 @@ from .exceptions import (
     ReasoningFormatError,
 )
 
-__version__ = "0.3.5"
+__version__ = "0.3.6"
 
 __all__ = [
     # Subpackages